move LineCollection to sub package
