+++ src/main/java/com/marginallyclever/makelangelo/donatelloimpl/nodes/turtle/LineWeightByImage.java
@@ -1,6 +1,6 @@
 package com.marginallyclever.makelangelo.donatelloimpl.nodes.turtle;
 
-import com.marginallyclever.convenience.LineCollection;
+import com.marginallyclever.convenience.linecollection.LineCollection;
 import com.marginallyclever.convenience.LineSegment2D;
 
 import com.marginallyclever.donatello.ports.InputDouble;
