+++ src/main/java/com/marginallyclever/makelangelo/makeart/turtlegenerator/lineweight/LineWeightByImageIntensity.java
@@ -1,6 +1,6 @@
 package com.marginallyclever.makelangelo.makeart.turtlegenerator.lineweight;
 
-import com.marginallyclever.convenience.LineCollection;
+import com.marginallyclever.convenience.linecollection.LineCollection;
 import com.marginallyclever.convenience.LineSegment2D;
 
 import com.marginallyclever.makelangelo.Translator;
